requirements.txt
streamlit>=1.32.0
google-genai>=1.0.0
numpy>=1.24.0
python-dotenv>=1.0.0